+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Use new filesystem API.
 * Pane enabled state is now remembered.
 * More efficient searches, especially for project scope.
+* Files and folders can be excluded from search.
 
 ## 0.3.0
 * Scope of search can be set to currently open document or entire project.